requirements.txt
duckdb>=1.0.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyarrow>=14.0.0
pytest>=8.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
openpyxl>=3.1.0
